+++ src/EventListener/LoadDataContainerListener.php
@@ -81,13 +81,11 @@ class LoadDataContainerListener
         }
 
         // Available for alerts settings
-        // if ($row['wemoffers_isAvailableForAlerts']) {
-        if ($row['isAlertCondition']) {
+        if (array_key_exists('isAlertCondition', $row) && $row['isAlertCondition']) {
             $data['eval']['isAlertCondition'] = true;
         }
 
         // Available for filters settings
-        // if ($row['wemoffers_isAvailableForFilters']) {
         if ($row['isFilter']) {
             $data['eval']['isFilter'] = true;
         }